Stop pacemaker and lustre via systemctl stop pacemaker
Rather than using pcs cluster kill, which is guaranteed to be
essentially instant but which doesn't stop lustre cleanly and
therefore risks filesystem corruption, stop pacemaker cleanly.

Recent tests have shown that is able to stop Lustre in about 5 minutes
on zinc and porter.

Past pacemaker versions shut down lustre and exported  pools on a node
but then started and stopped them on the partner sometimes.  This
issue appears to have been fixed.