ensure pacemaker is stopped and retry power state changes
After "timeout systemctl stop pacemaker", run pcs cluster kill to
handle the case where systemctl stop timed out.  We want to ensure
pacemaker is dead so it doesn't hold up the primary mgmt reboot.

After a power off or power on, wait a few seconds and check state.
If power state is not desired, try power off or on again.
It's sometimes the case that power state changes fail; it's very
unusual for them to fail twice if something isn't severely broken.