check pacemaker state before updating firmware
Update firmware is the point where we need the lustre servers to
be up and in a nominal state.  Before that all the action occurs
on the mgmt nodes.  Also, creating the image and updating the
mgmt nodes may take a long time, and lustre server state is more
likely to change the longer the period between the check and the
server updates and reboots