stop lustre with shutdown
Instead of running 'systemctl stop pacemaker' to unmount lustre on the servers,
pdsh 'shutdown' to the servers.  Perform the shutdown on OSS nodes first, then
on MDS nodes, as that seems to be faster.

When booting the nodes again, give more time before we start looping on checking
the hostlist.  An individual node takes about 2 minutes to get far into the boot
sequence, and a large set of nodes takes much longer.