allow more time for ib_burn_fw
When updating firmware on the Lustre servers, allow 300s for the process
to complete.

The timeout is in case one of the nodes simply crashes during the update.  Longer
term a better timeout mechanism would be good so that such situations are detected
and the script can fail immediately and allow the administrator to handle the issue.