Change pacemaker verification:
After updating primgmt node, exit if Pacemaker is not running and resources arent
up with instructions to sysadmin to fix issue, start pacemaker, and run lustre-mud
again.

Before bouncing sample node, exit if Pacemaker is not running with instructions to
sysadmin to fix issue, start pacemaker, and run lustre-mud again.  This covers
the case where pacemaker fails to start after primgmt is rebooted post tut update.

After bouncing sample node, exit if zpool fails to import or lustre target fails to
start.  Pacemaker does the import and the mount, so if that was successful we know
that pacemaker and the resource agents are at least mostly working.

After bouncing Lustre servers, verify that all resources are online.  If not,
exit with instructions to sysadmin to fix issue, start pacemaker, and verify the
update was successful.