include degraded pools when waiting for imports
When waiting for pools to be imported, we pdsh to the nodes and do
"zpool list".  We should count nodes with state DEGRADED as well as
pools with state ONLINE.  We often have pools resilvering after a drive
replacement, and this does not indicate a problem with the toss update.