update IB firmware differently in toss 3 and toss 4
The toss4 procedure is to install the mlxup and ib_firware packages
on the nodes before the reboot.

It's documented in Confluence under
LUSTRE/Updating+TOSS4+on+Lustre+Clusters

The toss3 procedure uses nfs mounts to make the firmware available
on the stateless nodes.

Either way the procedure assumes the primary mgmt node was updated
before starting the procedure on the server nodes.